+++ ProjetoViajeFacil/CSharp/ViajeFacilSolucao/ViajeFacil.Dominio.EF/Instituicao.cs
@@ -23,7 +23,7 @@ namespace ViajeFacil.Dominio.EF
 
         [Key]
         [Column(name: "id_instituicao")]
-        public int InstituicaoId { get; set; }
+        public long InstituicaoId { get; set; }
 
         [Column(name: "nome")]
         [Unicode(false)]
@@ -36,6 +36,6 @@ namespace ViajeFacil.Dominio.EF
         public string Telefone { get; set; } = null!;
 
         [Column(name: "id_endereco")]
-        public int EnderecoId { get; set; }
+        public long EnderecoId { get; set; }
     }
 }
